requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-admin-sortable2==2.1.3
django-appconf==1.0.5
django-ckeditor==6.5.1
django-debug-toolbar==3.6.0
django-imagekit==4.1.0
django-js-asset==2.0.0
pilkit==2.0
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
